libbladeRF: Fixed race condition in TX buffer submission
It is possible that the buffer completion callback occurs before
advance_tx_buffer() acts upon a successful return code from
async_submit_stream_buffer().

We must ensure the buffer is in the IN_FLIGHT state before
we temporarily drop the buffer lock (so a stream lock can be acquired),
or else the assertion at sync_worker.c:150 will not be upheld.

Therefore, we default the buffer state to being IN_FLIGHT before
dropping the lock, and revert it to FULL if the buffer submission
fails.